Renamed UIItemIdAppXmlConfiguration to IUIItemIdAppXmlConfiguration since it's an interface, Added default values for german
