+++ src/TestStack.White.UnitTests/Configuration/UIItemIdAppXmlConfigurationTests.cs
@@ -9,14 +9,14 @@ namespace TestStack.White.UnitTests.Configuration
         [Test]
         public void DefaultValues()
         {
-            Assert.That(UIItemIdAppXmlConfiguration.Instance.TableVerticalScrollBar, Is.EqualTo("Vertical Scroll Bar"));
-            Assert.That(UIItemIdAppXmlConfiguration.Instance.TableHorizontalScrollBar, Is.EqualTo("Horizontal Scroll Bar"));
-            Assert.That(UIItemIdAppXmlConfiguration.Instance.TableColumn, Is.EqualTo("Row "));
-            Assert.That(UIItemIdAppXmlConfiguration.Instance.TableTopLeftHeaderCell, Is.EqualTo("Top Left Header Cell"));
-            Assert.That(UIItemIdAppXmlConfiguration.Instance.TableCellNullValue, Is.EqualTo("(null)"));
-            Assert.That(UIItemIdAppXmlConfiguration.Instance.TableHeader, Is.EqualTo("Top Row"));
-            Assert.That(UIItemIdAppXmlConfiguration.Instance.HorizontalScrollBar, Is.EqualTo("Horizontal ScrollBar"));
-            Assert.That(UIItemIdAppXmlConfiguration.Instance.VerticalScrollBar, Is.EqualTo("Vertical ScrollBar"));
+            Assert.That(IUIItemIdAppXmlConfiguration.Instance.TableVerticalScrollBar, Is.EqualTo("Vertical Scroll Bar"));
+            Assert.That(IUIItemIdAppXmlConfiguration.Instance.TableHorizontalScrollBar, Is.EqualTo("Horizontal Scroll Bar"));
+            Assert.That(IUIItemIdAppXmlConfiguration.Instance.TableColumn, Is.EqualTo("Row "));
+            Assert.That(IUIItemIdAppXmlConfiguration.Instance.TableTopLeftHeaderCell, Is.EqualTo("Top Left Header Cell"));
+            Assert.That(IUIItemIdAppXmlConfiguration.Instance.TableCellNullValue, Is.EqualTo("(null)"));
+            Assert.That(IUIItemIdAppXmlConfiguration.Instance.TableHeader, Is.EqualTo("Top Row"));
+            Assert.That(IUIItemIdAppXmlConfiguration.Instance.HorizontalScrollBar, Is.EqualTo("Horizontal ScrollBar"));
+            Assert.That(IUIItemIdAppXmlConfiguration.Instance.VerticalScrollBar, Is.EqualTo("Vertical ScrollBar"));
         }
     }
 }
