+++ src/TestStack.White/UIItems/Scrolling/ScrollBars.cs
@@ -27,7 +27,7 @@ namespace TestStack.White.UIItems.Scrolling {
         protected virtual IHScrollBar FindHorizontalBar(FindElement findElement) {
             AutomationElement horizontalScrollElement =
                 findElement(
-                    AutomationSearchCondition.ByAutomationId(UIItemIdAppXmlConfiguration.Instance.HorizontalScrollBar).OfControlType(
+                    AutomationSearchCondition.ByAutomationId(IUIItemIdAppXmlConfiguration.Instance.HorizontalScrollBar).OfControlType(
                         ControlType.ScrollBar));
             if (horizontalScrollElement == null) {
                 return new NullHScrollBar();
@@ -42,7 +42,7 @@ namespace TestStack.White.UIItems.Scrolling {
         protected virtual IVScrollBar FindVerticalBar(FindElement findElement) {
             AutomationElement verticalScrollElement =
                 findElement(
-                    AutomationSearchCondition.ByAutomationId(UIItemIdAppXmlConfiguration.Instance.VerticalScrollBar).OfControlType(
+                    AutomationSearchCondition.ByAutomationId(IUIItemIdAppXmlConfiguration.Instance.VerticalScrollBar).OfControlType(
                         ControlType.ScrollBar));
             if (verticalScrollElement == null) {
                 return new NullVScrollBar();
